requirements.txt
# Core packages I installed manually
google-generativeai
langchain
langchain-community
langchain-google-genai
pypdf
chromadb
fastapi
uvicorn
python-docx
gradio
python-dotenv

# pip install -r requirements.txt